+++ temporian/core/test/registered_operators_test.py
@@ -95,6 +95,7 @@ class RegisteredOperatorsTest(absltest.TestCase):
             "SUBTRACTION",
             "SUBTRACTION_SCALAR",
             "TICK",
+            "TICK_CALENDAR",
             "TIMESTAMPS",
             "UNIQUE_TIMESTAMPS",
             "UNTIL_NEXT",